added interactive flag support

+++ src/rm/src/lib.rs
@@ -1,42 +1,66 @@
+use clap::ArgMatches;
+use promptly::prompt_default;
 use std::fs::{self, ReadDir};
 use std::path::Path;
-use clap::ArgMatches;
 
 #[allow(clippy::needless_lifetimes)]
-pub fn check_files<'a>(matches: &'a ArgMatches, paths: &'a Vec<&'a str>) -> anyhow::Result<Vec<&'a Path>> {
+pub fn check_files<'a>(
+    matches: &'a ArgMatches,
+    paths: &'a Vec<&'a str>,
+) -> anyhow::Result<Vec<&'a Path>> {
     let mut valid_paths = vec![];
 
-            for i in paths {
-                let path = Path::new(i);
-                if !path.exists() {
-                    panic!("file {:?} doesn't exist", path);
-                } else if path.is_dir() && !matches.is_present("recursive") {
-                        //"{:?} is a directory but -r modifier was not specified",
-                    return Err(anyhow::Error::msg(""));
-                } else {
-                    valid_paths.push(path)
-                }
-            }
-        
+    for i in paths {
+        let path = Path::new(i);
+        if !path.exists() {
+            panic!("file {:?} doesn't exist", path);
+        } else if path.is_dir() && !matches.is_present("recursive") {
+            //"{:?} is a directory but -r modifier was not specified",
+            return Err(anyhow::Error::msg(""));
+        } else {
+            valid_paths.push(path)
+        }
+    }
+
     Ok(valid_paths)
 }
 
 pub fn remove_files<'a>(matches: &'a ArgMatches, valid_paths: Vec<&'a Path>) -> anyhow::Result<()> {
     for path in valid_paths {
         if path.is_dir() && matches.is_present("recursive") {
-            let files = fs::read_dir(path).unwrap_or_else(|_| panic!("error while reading dir {}", path.as_os_str().to_str().unwrap()));
+            let files = fs::read_dir(path).unwrap_or_else(|_| {
+                panic!(
+                    "error while reading dir {}",
+                    path.as_os_str().to_str().unwrap()
+                )
+            });
+            if matches.is_present("interactive") {
+                let prompt =
+                    prompt_default(format!("remove dir {}?", path.to_str().unwrap()), false);
+                if !prompt.unwrap() {
+                    continue;
+                }
+            }
             recursive_search_remove(matches, files)?;
             fs::remove_dir(path)
                 .unwrap_or_else(|_| panic!("error while removing directory {:?}", path));
-                if matches.is_present("verbose") {
-                    println!("removed {}", path.to_str().unwrap())
-                }
+
+            if matches.is_present("verbose") {
+                println!("removed {}", path.to_str().unwrap())
+            }
         } else {
+            if matches.is_present("interactive") {
+                let prompt =
+                    prompt_default(format!("remove file {}?", path.to_str().unwrap()), false);
+                if !prompt.unwrap() {
+                    continue;
+                }
+            }
             fs::remove_file(path)
                 .unwrap_or_else(|_| panic!("error while removing file {:?}", path));
-                if matches.is_present("verbose") {
-                    println!("removed {}", path.to_str().unwrap())
-                }
+            if matches.is_present("verbose") {
+                println!("removed {}", path.to_str().unwrap())
+            }
         }
     }
 
@@ -48,15 +72,36 @@ fn recursive_search_remove(matches: &ArgMatches, read_dir: ReadDir) -> anyhow::R
         match i {
             Ok(entry) => {
                 if entry.path().is_dir() {
-                    recursive_search_remove(matches, fs::read_dir(entry.path().to_str().unwrap()).unwrap())?
+                    recursive_search_remove(
+                        matches,
+                        fs::read_dir(entry.path().to_str().unwrap()).unwrap(),
+                    )?;
+                    if matches.is_present("interactive") {
+                        let prompt = prompt_default(
+                            format!("remove file {}?", entry.file_name().to_str().unwrap()),
+                            false,
+                        );
+                        if !prompt.unwrap() {
+                            continue;
+                        }
+                    }
+                    fs::remove_dir(entry.path().to_str().unwrap())?;
                 } else if entry.path().is_file() {
+                    if matches.is_present("interactive") {
+                        let prompt = prompt_default(
+                            format!("remove file {}?", entry.file_name().to_str().unwrap()),
+                            false,
+                        );
+                        if !prompt.unwrap() {
+                            continue;
+                        }
+                    }
                     fs::remove_file(entry.path())?;
                     if matches.is_present("verbose") {
                         println!("removed {}", entry.file_name().to_str().unwrap())
                     }
                 }
-
-            },
+            }
             Err(_) => todo!(),
         }
     }
@@ -68,17 +113,20 @@ mod tests {
     use super::*;
     use std::fs::remove_dir;
 
-    use rand::distributions::{Alphanumeric, DistString};
     use clap::{Arg, Command};
-
+    use rand::distributions::{Alphanumeric, DistString};
 
     #[test]
     fn test_file() {
-        // test Clap App 
-        let matches = Command::new("test").get_matches();
+        // test Clap App
+        let matches = Command::new("test")
+            .arg(Arg::new("verbose"))
+            .arg(Arg::new("interactive"))
+            .get_matches();
 
         let string = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
-        fs::File::create(&string).unwrap_or_else(|_| panic!("Error while creating test file {}", &string));
+        fs::File::create(&string)
+            .unwrap_or_else(|_| panic!("Error while creating test file {}", &string));
 
         let path_vec = vec![string.as_str()];
 
@@ -95,12 +143,14 @@ mod tests {
     fn test_dir() {
         // test Clap App with necessary flags
         let matches = Command::new("test")
-        .arg(Arg::new("recursive")
-            .short('r'))
-        .get_matches_from(vec!["test", "-r"]);
+            .arg(Arg::new("verbose"))
+            .arg(Arg::new("interactive"))
+            .arg(Arg::new("recursive").short('r'))
+            .get_matches_from(vec!["test", "-r"]);
 
         let string = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
-        fs::create_dir(&string).unwrap_or_else(|_| panic!("Error while creating test dir {}", &string));
+        fs::create_dir(&string)
+            .unwrap_or_else(|_| panic!("Error while creating test dir {}", &string));
 
         let path_vec = vec![string.as_str()];
 
@@ -116,19 +166,25 @@ mod tests {
     #[test]
     #[should_panic]
     fn test_dir_no_flags() {
-        // test Clap App 
-        let matches = Command::new("test").arg(Arg::new("recursive")).get_matches_from(vec!["test"]);
+        // test Clap App
+        let matches = Command::new("test")
+            .arg(Arg::new("recursive"))
+            .arg(Arg::new("interactive"))
+            .arg(Arg::new("recursive"))
+            .get_matches_from(vec!["test"]);
 
         let string = Alphanumeric.sample_string(&mut rand::thread_rng(), 16);
-        fs::create_dir(&string).unwrap_or_else(|_| panic!("Error while creating test dir {}", &string));
+        fs::create_dir(&string)
+            .unwrap_or_else(|_| panic!("Error while creating test dir {}", &string));
 
         let path_vec = vec![string.as_str()];
 
         match check_files(&matches, &path_vec) {
-            Err(_) => {remove_dir(string).unwrap(); panic!("success, test panics and deletes test dir") },
-            Ok(_) => remove_dir(string).unwrap() // test fails, still deletes test dir,
+            Err(_) => {
+                remove_dir(string).unwrap();
+                panic!("success, test panics and deletes test dir")
+            }
+            Ok(_) => remove_dir(string).unwrap(), // test fails, still deletes test dir,
         }
     }
-
-
 }
